requirements.txt
cloudinary==1.39.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
python-dotenv==1.0.1
